+++ src/com/intellij/idea/plugin/hybris/project/compile/ProjectCompileUtil.kt
@@ -1,0 +1,36 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.intellij.idea.plugin.hybris.project.compile
+
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.openapi.vfs.LocalFileSystem
+import java.nio.file.Path
+
+object ProjectCompileUtil {
+
+    fun triggerRefreshGeneratedFiles(bootstrapDirectory: Path) {
+        val paths = listOf(
+            bootstrapDirectory.resolve(HybrisConstants.GEN_SRC_DIRECTORY),
+            bootstrapDirectory.resolve(HybrisConstants.PLATFORM_MODEL_CLASSES_DIRECTORY),
+            bootstrapDirectory.resolve(HybrisConstants.BIN_DIRECTORY).resolve(HybrisConstants.JAR_MODELS),
+        )
+
+        LocalFileSystem.getInstance().refreshNioFiles(paths, true, true, null)
+    }
+}