+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Clean `modelclasses` and `gensrc` directories before the code generation [#909](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/909)
 - Trigger file system change after the code generation [#910](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/910)
 - Trigger code generation even if there are modules of various types [#911](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/911)
+- Trigger file system change if code generation failed [#919](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/919)
 
 ### `Project Import` enhancements
 - Added autoconfiguration of the JRebel Facet for custom extensions [#869](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/869)